# dashboard/docs/generated/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "IT Global Dashboard API",
    "description": "IT Global 비즈니스 관리 대시보드를 위한 종합 API입니다.\n\n## 주요 기능\n- 프로젝트 관리 및 추적\n- 매출 및 재무 분석\n- 실시간 모니터링 및 메트릭\n- 사용자 관리 및 인증\n- 캐시 및 시스템 관리\n\n## 인증 방식\n이 API는 여러 인증 방법을 지원합니다:\n- 세션 기반 인증 (웹 인터페이스)\n- Bearer 토큰 인증 (API 접근)\n- Google OAuth 2.0 (웹 인터페이스)\n\n## 표준 응답 형식\n모든 API 응답은 다음과 같은 표준 형식을 따릅니다:\n```json\n{\n  \"success\": true|false,\n  \"data\": any,\n  \"error\": {...},\n  \"meta\": {...}\n}\n```\n\n## 버전 관리\nAPI는 URL 경로 기반 버전 관리를 사용합니다 (/api/v1/, /api/v2/ 등).",
    "version": "1.0.0",
    "contact": {
      "name": "IT Global Development Team",
      "email": "dev@itglobal.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development server"
    },
    {
      "url": "https://dashboard.itglobal.com",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/api/v1/projects": {
      "get": {
        "operationId": "projects_v1.list_projects",
        "summary": "프로젝트 목록 조회 (페이지네이션 지원)",
        "description": "summary: 프로젝트 목록 조회\ndescription: 페이지네이션과 필터링을 지원하는 프로젝트 목록 조회\nparameters:\nin: query\ntype: integer\nminimum: 1\ndefault: 1\ndescription: 페이지 번호\nin: query\ntype: integer\nminimum: 1\nmaximum: 100\ndefault: 20\ndescription: 페이지당 항목 수\nin: query\ntype: string\ndefault: created_at\ndescription: 정렬 필드\nin: query\ntype: string\nenum: [asc, desc]\ndefault: desc\ndescription: 정렬 순서\nin: query\ntype: string\nenum: [active, completed, cancelled, on_hold]\ndescription: 상태별 필터링\nin: query\ntype: string\ndescription: 프로젝트명 검색\nresponses:\n200:\ndescription: 프로젝트 목록 조회 성공\nschema:\n$ref: '#/definitions/StandardResponse'",
        "tags": [
          "Projects v1",
          "name: page",
          "name: limit",
          "name: sort",
          "name: order",
          "name: status",
          "name: search"
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResponse"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "찾을 수 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "서버 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "인증 필요",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "권한 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "sessionAuth": []
          }
        ]
      },
      "post": {
        "operationId": "projects_v1.create_project",
        "summary": "새 프로젝트 생성",
        "description": "summary: 프로젝트 생성\ndescription: 새로운 프로젝트를 생성합니다\nparameters:\nin: body\nrequired: true\nschema:\ntype: object\nproperties:\nname:\ntype: string\ndescription: 프로젝트명 (필수)\nexample: \"웹사이트 리뉴얼\"\ndescription:\ntype: string\ndescription: 프로젝트 설명\nexample: \"회사 홈페이지 전면 리뉴얼 프로젝트\"\nstart_date:\ntype: string\nformat: date\ndescription: 시작일 (필수)\nexample: \"2025-01-01\"\nend_date:\ntype: string\nformat: date\ndescription: 종료일\nexample: \"2025-06-30\"\nbudget:\ntype: number\ndescription: 예산\nexample: 50000000\nrequired:\nresponses:\n201:\ndescription: 프로젝트 생성 성공\n400:\ndescription: 입력 데이터 유효성 검사 실패\n409:\ndescription: 중복된 프로젝트명",
        "tags": [
          "Projects v1",
          "name: body",
          "name",
          "start_date"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectSchema"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResponse"
                }
              }
            }
          },
          "201": {
            "description": "생성됨",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResponse"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "찾을 수 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "서버 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "인증 필요",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "권한 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "sessionAuth": []
          }
        ]
      }
    },
...
    "/api/projects": {
      "post": {
        "operationId": "create_project",
        "summary": "새 프로젝트 생성 API",
        "description": "/api/projects 경로에서 새로운 데이터를 생성합니다.",
        "tags": [
          "Projects"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectSchema"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResponse"
                }
              }
            }
          },
          "201": {
            "description": "생성됨",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResponse"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "찾을 수 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "서버 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "인증 필요",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "권한 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "sessionAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "StandardResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "요청 성공 여부"
          },
          "data": {
            "description": "응답 데이터"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorDetails"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "required": [
          "success",
          "data",
          "error",
          "meta"
        ]
      },
      "ErrorResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/StandardResponse"
          },
          {
            "type": "object",
            "properties": {
              "success": {
                "type": "boolean",
                "example": false
              },
              "data": {
                "type": "null"
              }
            }
          }
        ]
      },
      "ErrorDetails": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "에러 코드"
          },
          "message": {
            "type": "string",
            "description": "에러 메시지"
          },
          "details": {
            "type": "object",
            "description": "추가 에러 정보"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "ResponseMeta": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "응답 시간"
          },
          "version": {
            "type": "string",
            "description": "API 버전"
          },
          "request_id": {
            "type": "string",
            "format": "uuid",
            "description": "요청 ID"
          }
        },
        "required": [
          "timestamp",
          "version",
          "request_id"
        ]
      }
    },
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "sessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session"
      }
    }
  },
...
}